+++ src/com/bridgelabz/userregistration/UserRegistration.java
@@ -1,0 +1,27 @@
+package com.bridgelabz.userregistration;
+
+import java.util.Scanner;
+
+public class UserRegistration {
+    public static void main(String[] args) {
+
+        System.out.println("Welcome to User Registration!");
+        Validation v = new Validation();
+        Scanner sc = new Scanner(System.in);
+        System.out.println("Please enter first name: ");
+        String f_name = sc.next();
+        v.nameValidation(f_name);
+        System.out.println("Please enter last name: ");
+        String l_name = sc.next();
+        v.nameValidation(l_name);
+        System.out.println("Please enter your Email-ID: ");
+        String email = sc.next();
+        v.emailValidation(email);
+        System.out.println("Please enter your 10 digit phone number: ");
+        String phone = sc.next();
+        v.phoneNumberValidation(phone);
+        System.out.println("Please enter your password(it should be of minimum length 8): ");
+        String password = sc.next();
+        v.passwordValidation(password);
+    }
+}
