+++ src/com/bridgelabz/userregistration/Validation.java
@@ -1,4 +1,4 @@
-package Regex;
+package com.bridgelabz.userregistration;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -24,8 +24,7 @@ public class Validation {
     }
     public void emailValidation(String email)
     {
-//        String regex = "^[abc]{1}(.*[A-Za-z0-9-_+.])?@{1}[A-Za-z0-9]{1,}.com{1}(.[a-z]*)?$";
-        String regex = "^[abc]{1}(.*[A-Za-z0-9-_+])(?(?=.*.)+[A-Za-z0-9-_+]{1,})?@{1}[A-Za-z0-9]{1,}.com{1}(.[a-z]*)?$";
+        String regex = "^[abc]{1}(.*[A-Za-z0-9-_+.])?@{1}[A-Za-z0-9]{1,}.com{1}(.[a-z]*)?$";
         Pattern p = Pattern.compile(regex);
         Matcher m = p.matcher(email);
         check(m.find());
